Feature: Update hardcoding
