+++ app/routes/_layout.post.tsx
@@ -1,439 +1,538 @@
-import { Form, NavLink, useActionData, useLoaderData, useNavigate, useSubmit } from '@remix-run/react';
-import { json } from '@remix-run/node';
-import { useState, useEffect } from 'react';
-import { Turnstile } from '@marsidev/react-turnstile';
-import type { ActionFunctionArgs, MetaFunction } from '@remix-run/node';
+import { useForm, type SubmitHandler, FormProvider, useFormContext, useWatch, useFieldArray, useFormState, FieldErrors } from "react-hook-form"
+import { useEffect, useState } from "react"
+import { z } from "zod";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { Form, json, NavLink, useActionData, useLoaderData } from "@remix-run/react";
+import UserExplanation from "~/components/SubmitFormComponents/UserExplanation";
+import ClearFormButton from "~/components/SubmitFormComponents/ClearFormButton";
+import { H3 } from "~/components/Headings";
+import TagSelectionBox from "~/components/SubmitFormComponents/TagSelectionBox";
+import { getTagsCounts, prisma } from "~/modules/db.server";
+import TagCreateBox from "~/components/SubmitFormComponents/TagCreateBox";
+import TagPreviewBox from "~/components/SubmitFormComponents/TagPreviewBox";
+import { Modal } from "~/components/Modal";
+import type { ActionFunctionArgs, MetaFunction } from "@remix-run/node";
+import { useSubmit } from "@remix-run/react";
+import { getClientIPAddress } from "remix-utils/get-client-ip-address";
+import { createEmbedding } from "~/modules/embedding.server";
+import { FaCopy } from "react-icons/fa";
+import { NodeHtmlMarkdown } from "node-html-markdown";
+import { commonMetaFunction } from "~/utils/commonMetafunction";
 
-import DynamicTextInput from '~/components/SubmitFormComponents/DynamicTextInput';
-import TagSelectionBox from '~/components/SubmitFormComponents/TagSelectionBox';
-import SituationInput from '~/components/SubmitFormComponents/SituationInput';
-import StaticTextInput from '~/components/SubmitFormComponents/StaticTextInput';
-import Preview from '~/components/SubmitFormComponents/Preview';
-import TagPreviewBox from '~/components/SubmitFormComponents/TagPreviewBox';
-import TagCreateBox from '~/components/SubmitFormComponents/TagCreateBox';
-import UserExplanation from '~/components/SubmitFormComponents/UserExplanation';
-import ValidationCheckBox from '~/components/SubmitFormComponents/ValidationCheckBox';
-import TextTypeSwitcher from '~/components/SubmitFormComponents/TextTypeSwitcher';
-import ClearFormButton from '~/components/SubmitFormComponents/ClearFormButton';
-import { Modal } from '~/components/Modal';
+const postFormSchema = z.object({
+    postCategory: z.enum(["misDeed", "goodDeed"]),
+    situations: z.object({
+        who: z.string().min(10, { message: "5W1H+Then状況説明>「誰が」は必須です" }),
+        what: z.string().min(9, { message: "5W1H+Then状況説明>「何を」は必須です" }),
+        when: z.string().min(8, { message: "5W1H+Then状況説明>「いつ」は必須です" }),
+        where: z.string().min(7, { message: "5W1H+Then状況説明>「どこで」は必須です" }),
+        why: z.string().min(6, { message: "5W1H+Then状況説明>「なぜ」は必須です" }),
+        how: z.string().min(5, { message: "5W1H+Then状況説明>「どうやって」は必須です" }),
+        // biome-ignore lint/suspicious/noThenProperty: <explanation>
+        then: z.string().min(4, { message: "5W1H+Then状況説明>「どうなったか」は必須です" }),
+        assumption: z.array(z.string()).optional(),
+    }),
+    reflection: z.array(z.string()).min(0, { message: "「健常行動ブレイクポイント」もしくは「なぜやってよかったのか」は最低一つ入力してください" }),
+    counterReflection: z.array(z.string()).min(0, { message: "「どうすればよかったか」もしくは「やらなかったらどうなっていたか」は最低一つ入力してください" }),
+    note: z.array(z.string()).optional(),
+    selectedTags: z.array(z.string()).optional(),
+    createdTags: z.array(z.string()).optional(),
+    title: z.array(z.string())
+    .superRefine((value, ctx) => {
+      if (value.some((v) => v.includes('#'))) {
+        ctx.addIssue({
+          code: z.ZodIssueCode.custom,
+          message: 'タイトルに「#」（ハッシュタグ）を含めることはできません。',
+        })
+      }
 
-import { prisma } from '~/modules/db.server';
-import { getClientIPAddress } from 'remix-utils/get-client-ip-address';
-import { createEmbedding } from '~/modules/embedding.server';
-import { commonMetaFunction } from '~/utils/commonMetafunction';
+      if (value.every((v) => v.length < 10)) {
+        ctx.addIssue({
+          code: z.ZodIssueCode.custom,
+          message: 'タイトルは必須です',
+        })
+      }
+    })
+});
 
-interface Tag {
-    tagName: string;
-    count: number;
-}
+type Inputs = z.infer<typeof postFormSchema>;
 
 export async function loader () {
-    const CFTurnstileSiteKey = process.env.CF_TURNSTILE_SITEKEY || "";
-    const tags = await prisma.dimTags.findMany({
-        select: {
-          tagName: true,
-          _count: {
-            select: { relPostTags: true },
-          },
-        },
-        orderBy: {
-          relPostTags: {
-            _count: "desc",
-          },
-        },
-      });
-      const allTagsOnlyForSearch: Tag[] = tags.map((tag) => {
-        return { tagName: tag.tagName, count: tag._count.relPostTags };
-      });
+  const CFTurnstileSiteKey = process.env.CF_TURNSTILE_SITEKEY || "";
+  const tags = await getTagsCounts();
 
-    return json({ CFTurnstileSiteKey,  allTagsOnlyForSearch });
+  return json({ CFTurnstileSiteKey,  tags });
 }
 
-export default function Component() {
-    
-    const [situationValues, setSituationValues] = useState<{ [key: string]: string }>({});
-    const [assumptionValues, setAssumptionValues] = useState<string[]>([]);
-    const [reflectionValues, setReflectionValues] = useState<string[]>([]);
-    const [counterFactualReflectionValues, setCounterFactualReflectionValues] = useState<string[]>([]);
-    const [noteValues, setNoteValues] = useState<string[]>([]);
-    const [titleValues, setTitleValues] = useState<string[]>([]);
-    const [selectedType, setSelectedType] = useState<string>('misDeed');
-    const [selectedTags, setSelectedTags] = useState<string[]>([]);
-    const [createdTags, setCreatedTags] = useState<string[]>([]);
-    const [isSubmitting, setIsSubmitting] = useState(false);
-    const [showErrorModal, setShowErrorModal] = useState(false);
-    const [showSuccessModal, setShowSuccessModal] = useState(false);
-    const actionData = useActionData<typeof action>();
-    const navigate = useNavigate();
-    const submit = useSubmit();
+export default function App() {
+  const { CFTurnstileSiteKey, tags } = useLoaderData<typeof loader>();
 
-    useEffect(() => {
-        if (actionData?.success === false) {
-          setShowErrorModal(true);
-          setIsSubmitting(false);
-        } else if (actionData?.success === true) {
-          setShowSuccessModal(true);
-          setTimeout(() => {
-            clearInputs();
-            'postId' in actionData && navigate(`/archives/${actionData.postId}`);
-          }, 3000);
-        }
-      }, [actionData, navigate]);
+  const [selectedTags, setSelectedTags] = useState<string[]>([]);
+  const [createdTags, setCreatedTags] = useState<string[]>([]);
 
-    const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        submit(event.currentTarget, { replace: true } );
-        setIsSubmitting(true);
+  const handleTagSelection = (tags: string[]) => {
+    setSelectedTags(tags);
+    window.localStorage.setItem('selectedTags', JSON.stringify(tags));
+    methods.setValue("selectedTags", tags);
+  }
+
+  const handleTagCreated = (tag: string) => {
+    setCreatedTags([...createdTags, tag]);
+    window.localStorage.setItem('createdTags', JSON.stringify([...createdTags, tag]));
+    methods.setValue("createdTags", [...createdTags, tag]);
+  }
+
+  const handleTagRemoved = (tag: string) => {
+    setCreatedTags(createdTags.filter((t) => t !== tag));
+    window.localStorage.setItem('createdTags', JSON.stringify(createdTags.filter((t) => t !== tag)));
+    methods.setValue("createdTags", createdTags.filter((t) => t !== tag));
+  }
+
+  
+  
+  const formId = "post-form";
+
+
+  const getStoredValues = (): Inputs => {
+    if (typeof window === "undefined") return { title: "", postCategory: "misDeed" };
+    const stored = window.localStorage.getItem(formId);
+    return stored ? JSON.parse(stored) : { title: "", postCategory: "misDeed" };
+  }
+
+  const methods = useForm({
+    defaultValues: getStoredValues(),
+    resolver: zodResolver(postFormSchema),
+  });
+
+  const clearInputs = () => {
+    methods.reset();
+    window.localStorage.removeItem(formId);
+    window.location.reload();
+  }
+  const submit = useSubmit();
+  const onSubmit: SubmitHandler<Inputs> = (data) => {
+    const formData = new FormData();
+    formData.append("_action", "firstSubmit");
+    for (const [key, value] of Object.entries(data)) {
+      formData.append(key, JSON.stringify(value));
     }
-    
-    useEffect(() => {
-        const savedSituationValue = window.localStorage.getItem('situationValue');
-        const initialSituationValue = JSON.parse(savedSituationValue || '{}');
-        setSituationValues(initialSituationValue);
-        const savedAssumptionValue = window.localStorage.getItem('assumptionValue');
-        const initialAssumptionValue = JSON.parse(savedAssumptionValue || '[]');
-        setAssumptionValues(initialAssumptionValue);
-        const savedReflectionValue = window.localStorage.getItem('reflectionValue');
-        const initialReflectionValue = JSON.parse(savedReflectionValue || '[]');
-        setReflectionValues(initialReflectionValue);
-        const savedCounterFactualReflectionValue = window.localStorage.getItem('counterFactualReflectionValue');
-        const initialCounterFactualReflectionValue = JSON.parse(savedCounterFactualReflectionValue || '[]');
-        setCounterFactualReflectionValues(initialCounterFactualReflectionValue);
-        const savedNoteValue = window.localStorage.getItem('noteValue');
-        const initialNoteValue = JSON.parse(savedNoteValue || '[]');
-        setNoteValues(initialNoteValue);
-        const savedTitleValue = window.localStorage.getItem('titleValue');
-        const initialTitleValue = JSON.parse(savedTitleValue || '[]');
-        setTitleValues(initialTitleValue);
-        const savedSelectedType = window.localStorage.getItem('selectedType');
-        const initialSelectedType = savedSelectedType || 'misDeed';
-        setSelectedType(initialSelectedType);
-        const savedSelectedTags = window.localStorage.getItem('selectedTags');
-        const initialSelectedTags = JSON.parse(savedSelectedTags || '[]');
-        setSelectedTags(initialSelectedTags);
-        const savedCreatedTags = window.localStorage.getItem('createdTags');
-        const initialCreatedTags = JSON.parse(savedCreatedTags || '[]');
-        setCreatedTags(initialCreatedTags);
+        submit(formData, {
+      method: "post",
+      action: "/post",
+    });
+  };  
 
-    }, []);
+  useEffect(() => {
+    setInterval(() => {
+      if (typeof window !== "undefined") {
+        window.localStorage.setItem(formId, JSON.stringify(methods.getValues()));
+      }
+    }, 5000);
+  }, [methods.getValues]);
 
-    const handleSituationChange = (values: { [key: string]: string }) => {
-        setSituationValues(values);
-        window.localStorage.setItem('situationValue', JSON.stringify(values));
-    };
-    const handleAssumptionChange = (values: string[]) => {
-        setAssumptionValues(values);
-        window.localStorage.setItem('assumptionValue', JSON.stringify(values));
-    };
-    const handleReflectionChange = (values: string[]) => {
-        setReflectionValues(values);
-        window.localStorage.setItem('reflectionValue', JSON.stringify(values));
-    };
-    const handleCounterFactualReflectionChange = (values: string[]) => {
-        setCounterFactualReflectionValues(values);
-        window.localStorage.setItem('counterFactualReflectionValue', JSON.stringify(values));
-    }
-    const handleNoteChange = (values: string[]) => {
-        setNoteValues(values);
-        window.localStorage.setItem('noteValue', JSON.stringify(values));
-    }
-    const handleTitleChange = (values: string[]) => {
-        setTitleValues(values);
-        window.localStorage.setItem('titleValue', JSON.stringify(values));
-    }
-    const handleToggle = (selectedType: string) => {
-        setSelectedType(selectedType);
-        window.localStorage.setItem('selectedType', selectedType);
-    }
+  const postCategory = methods.watch("postCategory");
+  const actionData = useActionData<typeof action>();
 
-    const handleTagSelection = (tags: string[]) => {
-        setSelectedTags(tags);
-        window.localStorage.setItem('selectedTags', JSON.stringify(tags));
-    }
-    useEffect(() => {
-        window.localStorage.setItem('createdTags', JSON.stringify(createdTags));
-    }, [createdTags]);
-
-    const handleTagCreated = (tag: string) => {
-    setCreatedTags((prevCreatedTags) => [...prevCreatedTags, tag]);
-    };
-
-    const handleTagRemoved = (tag: string) => {
-    setCreatedTags(createdTags.filter(createdTag => createdTag !== tag));
-    };
-
-    const [isValid, setIsValid] = useState(false);
-    const handleValidationResult = (result: boolean) => {
-    setIsValid(result);
-    };
-
-    const [isValidUser, setIsValidUser] = useState(false);
-
-    const handleTurnstileValidation = (isValid: boolean) => {
-        setIsValidUser(isValid);
-    };
-
-    const clearInputs = () => {
-        console.log("CLEARBUTTON")
-        setSituationValues({});
-        setAssumptionValues([]);
-        setReflectionValues([]);
-        setCounterFactualReflectionValues([]);
-        setNoteValues([]);
-        setTitleValues([]);
-        setSelectedTags([]);
-        setCreatedTags([]);
-        window.localStorage.clear()
-        setSelectedType('misDeed');
-    };
-
-    const { CFTurnstileSiteKey, allTagsOnlyForSearch } = useLoaderData<typeof loader>();
-    
-    
-    return (
+  return (
+    <>
     <div className="templateSubmitForm">
-        <Form method="post" onSubmit={handleSubmit}>
+    <FormProvider {...methods}> 
+      <Form method="post" onSubmit={methods.handleSubmit(onSubmit)}>
         <UserExplanation />
         <br/>
         <NavLink
             className="inline-block align-baseline font-bold text-sm text-info underline underline-offset-4"
             to="/freeStylePost"
         >自由投稿フォームに移動</NavLink>
-        <ClearLocalStorageButton clearInputs={clearInputs}/>
-        <TextTypeSwitcher onToggle={handleToggle} parentComponentStateValue={selectedType}/>
-        <SituationInput
-            onInputChange={handleSituationChange}
-            parentComponentStateValues={situationValues}
-            selectedType={selectedType}
-        />
+        <div className="flex justify-start mt-6">
+          <ClearFormButton clearInputs={clearInputs}/>
+        </div>
         <br/>
-        <br/>
-        <DynamicTextInput
-            description='書ききれなかった前提条件はありますか？'
-            onInputChange={handleAssumptionChange}
-            parentComponentStateValues={assumptionValues}
-        />
-        <br/>
-        <br/>
-        {selectedType === 'misDeed' ? (
+        <TextTypeSwitcher />
+        <SituationInput />
+        <DynamicTextInput description="書ききれなかった前提条件はありますか？" key="situations.assumption" />
+
+        {postCategory === "misDeed" ? (
             <>
             <StaticTextInput
-                row={3}
-                title='健常行動ブレイクポイント'
+                rowNumber={3}
+                title="健常行動ブレイクポイント"
                 placeholders={["友人の言動は冗談だという事に気が付く必要があった","会話の中で自分がされた時に困るようなフリは避けるべきである"]}
-                description='上で記述した状況がどのような点でアウトだったのかの説明です。 できる範囲で構わないので、なるべく理由は深堀りしてください。 「マナーだから」は理由としては認められません。 健常者エミュレータはマナー講師ではありません。一つずつ追加してください。3つ記入する必要はありません。'
-                onInputChange={handleReflectionChange}
-                parentComponentStateValues={reflectionValues}
+                description="上で記述した状況がどのような点でアウトだったのかの説明です。 できる範囲で構わないので、なるべく理由は深堀りしてください。 「マナーだから」は理由としては認められません。 健常者エミュレータはマナー講師ではありません。一つずつ追加してください。3つ記入する必要はありません。"
+                registerKey="reflection"
             />
             <StaticTextInput
-                row={3}
-                title='どうすればよかったか'
+                rowNumber={3}
+                title="どうすればよかったか"
                 placeholders={["冗談に対してただ笑うべきだった","詠ませた後もその句を大げさに褒めるなどして微妙な空気にさせないべきだった"]}
-                description='5W1H状況説明、健常行動ブレイクポイントを踏まえ、どのようにするべきだったかを提示します。'
-                onInputChange={handleCounterFactualReflectionChange}
-                parentComponentStateValues={counterFactualReflectionValues}
+                description="5W1H状説明、健常行動ブレイクポイントを踏まえ、どのようにするべきだったかを提示します。"
+                registerKey="counterReflection"
             />
             </>
         ) : (
             <>
             <StaticTextInput
-                row={3}
+                rowNumber={3}
                 title='なぜやってよかったのか'
                 placeholders={["一般的に料理とは手間のかかる作業であり、相手がかけてくれた手間に対して何らかの形で報いること、もしくは報いる意思を示すことは相手に対して敬意を表していることと等しい。","敬意はコミュニケーションに対して良い作用をもたらす"]}
                 description='上で記述した行動がなぜやってよかったのか、理由を説明します。できる範囲で構わないので、なるべく理由は深堀りしてください。なんとなくただ「よかった」は理由としては認められません。一つずつ追加してください。3つ記入する必要はありません。'
-                onInputChange={handleReflectionChange}
-                parentComponentStateValues={reflectionValues}
+                registerKey="reflection"
             />
             <StaticTextInput
-                row={3}
+                rowNumber={3}
                 title='やらなかったらどうなっていたか'
-                placeholders={["相手がかけた手間に対して敬意を払わないことは相手を無下に扱っていることと等しい。", "関係が改善されることはなく、状況が悪ければ破局に至っていたかもしれない"]}
-                description='仮に上で記述した行動を実行しなかった場合、どのような不利益が起こりうるか記述してください。推論の範囲内で構いません。'
-                onInputChange={handleCounterFactualReflectionChange}
-                parentComponentStateValues={counterFactualReflectionValues}
+                placeholders={["相手がかけた手間に対して敬意をわないことは相手を無下に扱っていることと等しい。", "関係が改善されることはなく、状況が悪ければ破局に至っていたかもしれない"]}
+                description='仮に上で記述した行動を実行しなかった場合、どのような不利益が起こりうるか記述してください。推論の範囲内で構わない。'
+                registerKey="counterReflection"
             />
             </>
         )}
         <StaticTextInput
-            row={3}
+            rowNumber={3}
             title='備考'
             description='書ききれなかったことを書きます'
-            placeholders={selectedType === "misDeed" ? ["友人が詠んだ句は「ため池や 水がいっぱい きれいだね」だった"] : ["舌が過度に肥えてしまい、コンビニ弁当が食べられなくなった。"]}
-            onInputChange={handleNoteChange}
-            parentComponentStateValues={noteValues}
+            placeholders={postCategory === "misDeed" ? ["友人が詠んだ句は「ため池や 水がいっぱい きれいだね」だった"] : ["舌が過度に肥えてしまい、コンビニ弁当が食べられなくなった。"]}
+            registerKey="note"
         />
-        <Preview
-            selectedType={selectedType}
-            situationValues={situationValues}
-            assumptionValues={assumptionValues}
-            reflectionValues={reflectionValues}
-            counterFactualReflectionValues={counterFactualReflectionValues}
-            noteValues={noteValues}
-        />
-        <br/>
-        <StaticTextInput
-            row={1}
-            title='タイトル'
-            description='得られる教訓を要約してください'
-            onInputChange={handleTitleChange}
-            parentComponentStateValues={titleValues}
-        />
-        <TagSelectionBox
-            onTagsSelected={handleTagSelection}
-            parentComponentStateValues={selectedTags}
-            allTagsOnlyForSearch={allTagsOnlyForSearch}
-        />
-        <br/>
+        <TagSelectionBox allTagsOnlyForSearch={tags} onTagsSelected={handleTagSelection} parentComponentStateValues={selectedTags} />
         <TagCreateBox
             handleTagCreated={handleTagCreated}
             handleTagRemoved={handleTagRemoved}
             parentComponentStateValues={createdTags}
         /> 
         <TagPreviewBox selectedTags={selectedTags} createdTags={createdTags}/>
-        <ValidationCheckBox
-            titleValues={titleValues}
-            situationValues={situationValues}
-            reflectionValues={reflectionValues}
-            counterFactualReflectionValues={counterFactualReflectionValues}
-            onValidationResult={handleValidationResult}
+        <StaticTextInput
+            rowNumber={1}
+            title="タイトル"
+            description="タイトルを入力してください"
+            placeholders={["タイトル"]}
+            registerKey="title"
         />
-        <Turnstile siteKey={CFTurnstileSiteKey} onSuccess={() => handleTurnstileValidation(true)}/>
-        <button
-            type="submit"
-            className={`rounded-md block w-full px-4 py-2 text-center my-4 ${
-              isValid && isValidUser && !isSubmitting
-                ? 'btn-primary'
-                : 'bg-gray-400 cursor-not-allowed text-white'
-            }`}
-            disabled={!isValid || !isValidUser || isSubmitting}
-        >
-        {isSubmitting ? '投稿中...' : '投稿する'}
-        </button>
-        <input type="hidden" name="situationValues" value={JSON.stringify(situationValues)} />
-        <input type="hidden" name="assumptionValues" value={JSON.stringify(assumptionValues)} />
-        <input type="hidden" name="reflectionValues" value={JSON.stringify(reflectionValues)} />
-        <input type="hidden" name="counterFactualReflectionValues" value={JSON.stringify(counterFactualReflectionValues)} />
-        <input type="hidden" name="noteValues" value={JSON.stringify(noteValues)} />
-        <input type="hidden" name="titleValues" value={JSON.stringify(titleValues)} />
-        <input type="hidden" name="selectedType" value={selectedType} />
-        <input type="hidden" name="selectedTags" value={JSON.stringify(selectedTags)} />
-        <input type="hidden" name="createdTags" value={JSON.stringify(createdTags)} />
-        </Form>
-        <Modal
-            isOpen={showErrorModal}
-            onClose={() => setShowErrorModal(false)}
-            title="エラー"
-        >
-            <p>{actionData?.error}</p>
-        </Modal>
-
-        <Modal
-            isOpen={showSuccessModal}
-            onClose={() => setShowSuccessModal(false)}
-            title="投稿成功"
-            showCloseButton={false}
-        >
-            <p>投稿が成功しました。リダイレクトします...</p>
-        </Modal>
+        <PreviewButton data={actionData?.data} />
+      </Form>
+      </FormProvider>
     </div>
+    </>
+  )
+}
+
+
+
+
+function TextTypeSwitcher(){
+    const { register } = useFormContext();
+    return (
+        <div className="mb-4">
+            <H3>投稿タイプを選択</H3>
+            <p>投稿したい経験知の種類を選択してください。</p>
+            <div className="flex mt-4">
+                <input type="radio" id="misDeed" value="misDeed" {...register("postCategory")} />
+                <label htmlFor="misDeed">結果悪</label>
+                <input type="radio" id="goodDeed" value="goodDeed" {...register("postCategory")} />
+                <label htmlFor="goodDeed">結果善</label>
+            </div>
+        </div>
     )
 }
 
-function removeEmptyString(array: string[]): string[] {
-    return array.filter((value) => !/^\s*$/.test(value));
+
+function SituationInput(){
+    const { control, register, formState: { errors } } = useFormContext();
+    const postCategory = useWatch({ control, name: "postCategory" });
+    const placeholder = postCategory === "misDeed" ? [
+        {
+            key: "who",
+            description: "その状況の「主役」は誰ですか？(Who)",
+            placeholder: "自分が",
+            rows: 1,
+          },
+          {
+            key: "when",
+            description: "いつ起こったことですか？(When)",
+            placeholder: "友人と公園にいた時",
+            rows: 1,
+          },
+          {
+            key: "where",
+            description: "どこで起こったことですか？(Where)",
+            placeholder: "池の前で",
+            rows: 1,
+          },
+          {
+            key: "why",
+            description: "なぜそのような行動をしたのですか？(Why)",
+            placeholder:
+              "「詠めそう」と言われたらそう返すのが自然な会話の流れだと思ったから",
+            rows: 2,
+          },
+          {
+            key: "what",
+            description: "その主役は、何に対してはたらきかけましたか？(What)",
+            placeholder: "友人に",
+            rows: 1,
+          },
+          {
+            key: "how",
+            description: "その主役は、対象をどうしましたか？(How)",
+            placeholder: "「詠んでみてよ」と言った",
+            rows: 1,
+          },
+          {
+            key: "then",
+            description: "行動の結果としてどうなりましたか？(Then)",
+            placeholder: "友人が微妙な句を詠み、微妙な空気になった",
+            rows: 3,
+          },
+        ]
+      : [
+          {
+            key: "who",
+            description: "その状況の「主役」は誰ですか？(Who)",
+            placeholder: "筆者が",
+            rows: 1,
+          },
+          {
+            key: "when",
+            description: "いつ起こったことですか？(When)",
+            placeholder: "コロナで同居人が家にいる時間が増えた時",
+            rows: 1,
+          },
+          {
+            key: "where",
+            description: "どこで起こったことですか？(Where)",
+            placeholder: "家で",
+            rows: 1,
+          },
+          {
+            key: "why",
+            description: "なぜそのような行動をしたのですか？(Why)",
+            placeholder: "おいしいと思ったため",
+            rows: 2,
+          },
+          {
+            key: "what",
+            description: "その主役は、何に対してはたらきかけましたか？(What)",
+            placeholder: "同居人が作ってくれる料理について",
+            rows: 1,
+          },
+          {
+            key: "how",
+            description: "その主役は、対象をどうしましたか？(How)",
+            placeholder: "相手に直接「おいしい」と伝えるようにした",
+            rows: 1,
+          },
+          {
+            key: "then",
+            description: "行動の結果としてどうなりましたか？(Then)",
+            placeholder:
+              "相手の料理の腕が上がり、どんどん料理がおいしくなり、関係も改善された",
+            rows: 3,
+          },
+        ];
+
+
+    return (
+        <div className="mb-8">
+            <H3>5W1H+Then状況説明</H3>
+            {placeholder.map((data, index) => {
+                return (
+                    <div key={data.key} className="mb-4">
+                        <label
+                            htmlFor={data.key}
+                            className="block font-bold mb-2"
+                        >
+                            {data.description}
+                        </label>
+                        <textarea
+                            id={data.key}
+                            placeholder={data.placeholder}
+                            rows={data.rows}
+                            className="w-full px-3 py-2 text-base-content border rounded-lg focus:outline-none placeholder-slate-500"
+                            {...register(`situations.${data.key}`)}
+                        />
+                        <div>
+                        {errors.situations?.[data.key] && <ErrorMessageContainer errormessage={errors.situations?.[data.key]?.message as string} />}
+                        </div>
+                    </div>
+                )
+            })}
+        </div>
+    )
+}
+
+function DynamicTextInput({ description, registerKey = "situations.assumption" }: {description: string, registerKey?: string}){
+    const { register, control } = useFormContext();
+    const { fields, append, remove } = useFieldArray({
+        control,
+        name: registerKey,
+    });
+    return (
+      <div className="mb-8">
+        <H3>{description}</H3>
+        <div className="flex mb-4 flex-col">
+          {fields.map((field, index) => (
+            <div key={field.id} className="flex flex-row my-2">
+              <input key={field.id} {...register(`${registerKey}.${index}`)} className="w-3/4 border rounded-lg placeholder-slate-500 px-3 py-2" />
+              <button type="button" onClick={() => remove(index)} className="bg-error text-error-content rounded-lg px-3 mx-3">削除</button>
+            </div>
+          ))}
+        </div>
+        <button type="button" onClick={() => append("")} className="bg-info text-info-content px-3 py-2 rounded-lg">追加</button>
+      </div>
+    )
+}
+
+function StaticTextInput({ rowNumber, title, placeholders, description, registerKey }: { rowNumber: number, title: string, placeholders: string[], description: string, registerKey: string }){
+    const { register, formState: { errors }
+   } = useFormContext();
+    const renderTextInputs = () => {
+        const inputs = [];
+        for (let i = 0; i < rowNumber; i++) {
+            inputs.push(
+              <div className="w-full">
+                <textarea
+                    key={`${title}-input-${i}`}
+                    className="flex-grow px-3 py-2 border rounded-lg focus:outline-none w-full my-2 placeholder-slate-500"
+                    placeholder={placeholders[i]}
+                    {...register(`${registerKey}.${i}`)}
+                />
+                {errors[registerKey] && <ErrorMessageContainer errormessage={errors[registerKey]?.root?.message as string} />}
+              </div>
+            )
+        }
+        return inputs;
+    }
+    return (
+        <div className="mb-8">
+            <H3>{title}</H3>
+            <p>{description}</p>
+            <div className="flex items-center space-x-2 mb-4 flex-col">
+                {renderTextInputs()}             
+            </div>
+        </div>
+    )
+}
+
+function ErrorMessageContainer({errormessage}: {errormessage: string}){
+  return (
+    <div>
+      <p className="text-error">
+        [!] {errormessage}
+      </p>
+    </div>
+  )
+}
+
+
+function PreviewButton({ data }: { data?: { WikifiedResult: string, MarkdownResult: string } }){
+  const [showPreviewModal, setShowPreviewModal] = useState(false);
+  const { getValues } = useFormContext();
+  const submit = useSubmit();
+
+  const handleSubmit = () => {
+    const data = getValues() as Inputs;
+    const formData = new FormData();
+    formData.append("_action", "secondSubmit");
+    for (const [key, value] of Object.entries(data)) {
+      formData.append(key, JSON.stringify(value));
+    }
+    submit(formData, { method: "post", action: "/post" });
   }
 
-export async function action({ request }:ActionFunctionArgs ) {
-    const formData = await request.formData();
-    const situationValues = JSON.parse(formData.get('situationValues') as string);
-    let assumptionValues = JSON.parse(formData.get('assumptionValues') as string);
-    let reflectionValues = JSON.parse(formData.get('reflectionValues') as string);
-    let counterFactualReflectionValues = JSON.parse(formData.get('counterFactualReflectionValues') as string);
-    let noteValues = JSON.parse(formData.get('noteValues') as string);
-    const titleValues = JSON.parse(formData.get('titleValues') as string);
-    const selectedType = formData.get('selectedType') as string;
-    const selectedTags = JSON.parse(formData.get('selectedTags') as string);
-    const createdTags = JSON.parse(formData.get('createdTags') as string);
-    
-    assumptionValues = removeEmptyString(assumptionValues);
-    reflectionValues = removeEmptyString(reflectionValues);
-    counterFactualReflectionValues = removeEmptyString(counterFactualReflectionValues);
-    noteValues = removeEmptyString(noteValues);
+  const handleCopy = async () => {
+    if (data?.MarkdownResult){
+      navigator.clipboard.writeText(data.MarkdownResult);
+    }
+  }
 
-    if (titleValues.some((title: string) => title.includes('#'))){
-        return json({
-            success: false,
-            error: 'タイトルに「#」（ハッシュタグ）を含めることはできません。'
-        });
+
+  return (
+    <div className="flex justify-end">
+      <button type="submit" onClick={() => setShowPreviewModal(true)} className="btn btn-primary">投稿する</button>
+      <Modal
+        isOpen={showPreviewModal}
+        onClose={() => setShowPreviewModal(false)}
+        title="投稿内容を確認してください"
+        showCloseButton={false}
+      >
+        <div className="postContent">
+          {/* biome-ignore lint/security/noDangerouslySetInnerHtml: <explanation> */}
+          <div dangerouslySetInnerHTML={{ __html: data?.WikifiedResult ?? "" }} />
+        </div>
+        <div className="flex justify-between mt-6">
+          <button type="button" onClick={() => setShowPreviewModal(false)} className="btn btn-secondary">修正する</button>
+          <button type="button" onClick={handleCopy} className="btn btn-secondary"><FaCopy /></button>
+          <button type="button" onClick={handleSubmit} className="btn btn-primary">投稿する</button>
+        </div>
+      </Modal>
+    </div>
+  )
+}
+
+
+
+export async function action({ request }:ActionFunctionArgs){
+  const formData = await request.formData();
+  const actionType = formData.get("_action");
+  const postData = Object.fromEntries(formData);
+  const parsedData = {
+    ...postData,
+    postCategory: JSON.parse(postData.postCategory as string),
+    situations: JSON.parse(postData.situations as string),
+    reflection: JSON.parse(postData.reflection as string),
+    counterReflection: JSON.parse(postData.counterReflection as string),
+    note: JSON.parse(postData.note as string),
+    selectedTags: JSON.parse(postData.selectedTags as string),
+    title: JSON.parse(postData.title as string),
+  } as unknown as Inputs;
+
+  if (actionType === "firstSubmit"){
+    try{  
+      const html = await Wikify(parsedData);
+      const data = await html.json();
+      return json({
+        success: "success",
+        data: data.data,
+        error: undefined,
+      });
+    } catch (error) {
+      return json({
+        success: "failed",
+        data: undefined,
+        error: "Wikify failed",
+      });
+    }
+  }
+
+  if (actionType === "secondSubmit"){
+    const html = await Wikify(parsedData);
+    const data = await html.json();
+    const isSuccess = data.success;
+    const wikifyResult = data.data?.WikifiedResult;
+    const postTitle = parsedData.title[0];
+    const createdTags = parsedData.createdTags;
+    const selectedTags = parsedData.selectedTags;
+    if (!wikifyResult) {
+      return json({ success: false, error: "Wikify failed", data: undefined });
     }
 
-    const result = `
-        <h3>5W1H+Then状況説明</h3>
-        <table><tbody>
-          <tr><td>Who(誰が)</td><td>${situationValues.who || ''}</td></tr>
-          <tr><td>When(いつ)</td><td>${situationValues.when || ''}</td></tr>
-          <tr><td>Where(どこで)</td><td>${situationValues.where || ''}</td></tr>
-          <tr><td>Why(なぜ)</td><td>${situationValues.why || ''}</td></tr>
-          <tr><td>What(何を)</td><td>${situationValues.what || ''}</td></tr>
-          <tr><td>How(どのように)</td><td>${situationValues.how || ''}</td></tr>
-          <tr><td>Then(どうした)</td><td>${situationValues.then || ''}</td></tr>
-        </tbody></table>
-        ${assumptionValues.length > 0 ? `
-          <h3>前提条件</h3>
-          <ul>
-            ${assumptionValues.map((assumption: string) => `<li>${assumption}</li>`).join('')}
-          </ul>
-        ` : ''}
-        <h3>
-          ${selectedType === 'misDeed'
-            ? '健常行動ブレイクポイント'
-            : selectedType === 'goodDeed'
-            ? 'なぜやってよかったのか'
-            : ''}
-        </h3>
-        <ul>
-          ${reflectionValues.map((reflection: string) => `<li>${reflection}</li>`).join('')}
-        </ul>
-
-        <h3>
-          ${selectedType === 'misDeed'
-            ? 'どうすればよかったか'
-            : selectedType === 'goodDeed'
-            ? 'やらなかったらどうなっていたか'
-            : ''}
-        </h3>
-        <ul>
-          ${counterFactualReflectionValues.map((counterFactualReflection: string) => `<li>${counterFactualReflection}</li>`).join('')}
-        </ul>
-
-        ${noteValues.length > 0 ? `
-          <h3>備考</h3>
-          <ul>
-            ${noteValues.map((note: string) => `<li>${note}</li>`).join('')}
-          </ul>
-        ` : ''}
-      `;
-
-    const ip = getClientIPAddress(request) || "";
-    const postUserIpHash = await crypto.subtle.digest(
-        "SHA-256",
-        new TextEncoder().encode(ip)
-    );
-    const hashArray = Array.from(new Uint8Array(postUserIpHash));
-    const postUserIpHashString = hashArray.map((b) => b.toString(16).padStart(2, "0")).join("");
-
-    const newPost = await prisma.$transaction(async (prisma) => {
+    if (isSuccess){
+      const hashedUserIpAddress = await getHashedUserIPAddress(request);
+      if (data.data === undefined){
+        return json({ success: false, error: "Wikify failed", data: undefined });
+      }
+      const newPost = await prisma.$transaction(async (prisma) => {
         const newPost = await prisma.dimPosts.create({
-            data: {
-                postAuthorIPHash: postUserIpHashString,
-                postContent: result,
-                postTitle: titleValues.join(''),
-                countLikes: 0,
-                countDislikes: 0,
-                commentStatus: "open",
-            },
+          data: {
+            postAuthorIPHash: hashedUserIpAddress,
+            postContent: wikifyResult,
+            postTitle: postTitle,
+            countLikes: 0,
+            countDislikes: 0,
+            commentStatus: "open",
+        },
         });
-        const uniqueTags = [...new Set([...selectedTags, ...createdTags])];
+        const uniqueTags = [...new Set([...selectedTags || [], ...createdTags || []])];
         const existingTags = await prisma.dimTags.findMany({
             where: {
                 tagName: {
@@ -441,53 +540,103 @@ export async function action({ request }:ActionFunctionArgs ) {
                 },
             },
         });
-
         const existingTagNames = existingTags.map((tag) => tag.tagName);
         const newTagNames = uniqueTags.filter((tag) => !existingTagNames.includes(tag));
-
-        const newTags = await Promise.all(
-            newTagNames.map((tagName) =>
-                prisma.dimTags.create({
-                    data: {
-                        tagName,
-                    },
-                })
-            )
-        );
-
+        const newTags = await Promise.all(newTagNames.map(async (tagName) => {
+          return await prisma.dimTags.create({ data: { tagName } });
+        }));
         const allTags = [...existingTags, ...newTags];
-        await Promise.all(
-            allTags.map((tag) =>
-                prisma.relPostTags.create({
-                    data: {
-                        postId: newPost.postId,
-                        tagId: tag.tagId,
-                    },
-                })
-            )
-        );
-        
-        return newPost
-    });
+        await Promise.all(allTags.map(async (tag) => {
+          return await prisma.relPostTags.create({ data: { postId: newPost.postId, tagId: tag.tagId } });
+        }));
+        return newPost;
+      });
 
-    await createEmbedding({
+      await createEmbedding({
         postId: Number(newPost.postId),
         postContent: newPost.postContent,
         postTitle: newPost.postTitle,
-    });
+      });
+      return json({ success: true, error: undefined, data: newPost });
+    }
+    return json({ success: false, error: data.error, data: undefined });
+  }
+}
 
-    return json(
-        { success: true, postId: newPost.postId, error: null }
-    );
+async function getHashedUserIPAddress(request: Request){
+  const ipAddress = getClientIPAddress(request) || "";
+  const postUserIpHash = await crypto.subtle.digest(
+    "SHA-256",
+    new TextEncoder().encode(ipAddress)
+  );
+  const hashArray = Array.from(new Uint8Array(postUserIpHash));
+  const postUserIpHashString = hashArray.map((b) => b.toString(16).padStart(2, "0")).join("");
+  return postUserIpHashString;
+}
+
+async function Wikify(postData: Inputs){
+  // バリデーションを実施
+  const validationResult = postFormSchema.safeParse(postData);
+  if (!validationResult.success) {
+    return json({
+      success: false,
+      error: validationResult.error.errors,
+      data: undefined,
+    });
+  }
+
+  const { who, when, where, why, what, how, then, assumption } = validationResult.data.situations;
+  const { reflection, counterReflection } = validationResult.data;
+  const { title, note, selectedTags, createdTags, postCategory } = validationResult.data;
+
+  function removeEmptyString(array: string[]): string[] {
+    return array.filter((value) => !/^\s*$/.test(value));
+  }
+
+  const result = `
+    <h3>5W1H+Then状況説明</h3>
+    <table><tbody>
+      <tr><td>Who(誰が)</td><td>${who}</td></tr>
+      <tr><td>When(いつ)</td><td>${when}</td></tr>
+      <tr><td>Where(どこで)</td><td>${where}</td></tr>
+      <tr><td>Why(なぜ)</td><td>${why}</td></tr>
+      <tr><td>What(何を)</td><td>${what}</td></tr>
+      <tr><td>How(どのように)</td><td>${how}</td></tr>
+      <tr><td>Then(どうした)</td><td>${then}</td></tr>
+    </tbody></table>
+    ${assumption && assumption.length > 0 ? `
+      <h3>前提条件</h3>
+      <ul>
+        ${removeEmptyString(assumption)?.map((assumption) => `<li>${assumption}</li>`).join('\n')}
+      </ul>
+      ` : ''}
+    <h3>
+      ${postCategory === "misDeed" ? "健常行動ブレイクポイント" : postCategory === "goodDeed" ? "なぜやってよかったのか" : ""}
+    </h3>
+    <ul>
+      ${removeEmptyString(reflection)?.map((reflection) => `<li>${reflection}</li>`).join('\n')}
+    </ul>
+    <h3>
+      ${postCategory === "misDeed" ? "どうすればよかったか" : postCategory === "goodDeed" ? "やらなかったらどうなっていたか" : ""}
+    </h3>
+    <ul>
+      ${removeEmptyString(counterReflection)?.map((counterReflection) => `<li>${counterReflection}</li>`).join('\n')}
+    </ul>
+    ${note && note.length > 0 ? `
+      <h3>備考</h3>
+      ${removeEmptyString(note)?.map((note) => `<li>${note}</li>`).join('\n')}
+    ` : ''}
+  `
+  const markdownContent = NodeHtmlMarkdown.translate(result);
+  return json({ success: true, error: undefined, data: { WikifiedResult : result, MarkdownResult : markdownContent } });
 }
 
 export const meta: MetaFunction = () => {
-    const commonMeta = commonMetaFunction({
-        title : "投稿する",
-        description : "テンプレートに沿って投稿する",
-        url: "https://healthy-person-emulator.org/post",
-        image: null
-    });
-    return commonMeta;
+  const commonMeta = commonMetaFunction({
+      title : "投稿する",
+      description : "テンプレートに沿って投稿する",
+      url: "https://healthy-person-emulator.org/post",
+      image: null
+  });
+  return commonMeta;
 };
-
